Deprecate dependency map notation
TODO: Add upgrade guide section

In order to avoid the notion that Gradle always has more than one way to do things, we have decided
to deprecate the map notation for dependencies. It is the more verbose compared to the string notation
and is arguably less discoverable as it relies on map entries.

TODO: Add tests verifying deprecation status